fix: compat utils
