+++ catalyst-gateway/bin/src/service/api/documents/put_document/mod.rs
@@ -15,9 +15,12 @@ use crate::{
         },
         index::session::CassandraSessionError,
     },
-    service::common::{
-        auth::rbac::token::CatalystRBACTokenV1, responses::WithErrorResponses,
-        types::headers::retry_after::RetryAfterOption,
+    service::{
+        api::documents::common::compat,
+        common::{
+            auth::rbac::token::CatalystRBACTokenV1, responses::WithErrorResponses,
+            types::headers::retry_after::RetryAfterOption,
+        },
     },
 };
 
@@ -198,6 +201,8 @@ async fn store_document_in_db(
         authors,
         Some(doc_meta_json),
     );
+
+    let doc_body = compat::to_new_version(doc_body)?;
 
     FullSignedDoc::new(doc_body, payload, doc_bytes)
         .store()
